Moved more of the schemaEditor logic from ./shared into the SchemaEditor in order to place Schema inspector next to the editor and toolbar. Refactored some imports.
