+++ src/studio/src/designer/frontend/packages/schema-editor/package.json
@@ -30,13 +30,16 @@
     "@testing-library/jest-dom": "~5.14.1",
     "@testing-library/react": "~12.0.0",
     "@testing-library/user-event": "~12.8.3",
+    "@types/enzyme": "~3.10.9",
     "@types/jest": "~26.0.24",
     "@types/react": "~17.0.14",
     "@types/react-dom": "~17.0.9",
     "@types/react-redux": "~7.1.18",
+    "@types/redux-mock-store": "~1.0.3",
     "axios": "~0.21.1",
     "cross-env": "~7.0.3",
     "css-loader": "~5.2.6",
+    "enzyme": "~3.11.0",
     "html-loader": "~1.3.2",
     "html-webpack-plugin": "~4.5.2",
     "jest": "~27.0.6",
@@ -50,6 +53,7 @@
     "react-redux": "~7.2.4",
     "react-scripts": "4.0.3",
     "react-styleguidist": "~11.1.7",
+    "redux-mock-store": "~1.5.4",
     "rollup": "~2.53.3",
     "rollup-plugin-css-only": "~3.1.0",
     "rollup-plugin-peer-deps-external": "~2.2.4",
